refactor(vm): Rename the serializer module
