+++ vm/src/api/ser.rs
